# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: OVELIN store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/support/tickets:
    get:
      operationId: listAdminSupportTickets
      tags: [admin]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AdminSupportTicket" } }
...
  /admin/support/tickets/{id}:
    get:
      operationId: getAdminSupportTicket
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Ticket detail
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminSupportTicketDetail" } } }
    patch:
      operationId: updateAdminSupportTicket
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateAdminSupportTicketBody" } } }
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/SuccessResponse" } } }
  /admin/support/tickets/{id}/reply:
    post:
      operationId: replyAdminSupportTicket
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SupportMessageBody" } } }
      responses:
        "200":
          description: Replied
          content: { application/json: { schema: { $ref: "#/components/schemas/SuccessResponse" } } }
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AdminLoginBody" } } }
      responses:
        "200":
          description: Admin session
          content: { application/json: { schema: { $ref: "#/components/schemas/SuccessResponse" } } }
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      responses:
        "200":
          description: Settings
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminSettings" } } }
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateSettingsBody" } } }
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/AdminSettings" } } }
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }
...
    AdminSettings:
      type: object
      required: [siteName, supportWhatsapp, supportTelegram, depositUsdtAddress, depositBankInfo, depositCashInfo, minWithdraw, referralCommissionPct, referralSignupBonus, cashbackPct, maintenanceMode, maintenanceMessage, announcementBar, vipBronzeMin, vipSilverMin, vipGoldMin, vipDiamondMin]
      properties:
        siteName: { type: string }
        supportWhatsapp: { type: string }
        supportTelegram: { type: string }
        depositUsdtAddress: { type: string }
        depositBankInfo: { type: string }
        depositCashInfo: { type: string }
        minWithdraw: { type: string }
        referralCommissionPct: { type: string }
        referralSignupBonus: { type: string }
        cashbackPct: { type: string }
        maintenanceMode: { type: boolean }
        maintenanceMessage: { type: string }
        announcementBar: { type: string }
        vipBronzeMin: { type: string }
        vipSilverMin: { type: string }
        vipGoldMin: { type: string }
        vipDiamondMin: { type: string }
    UpdateSettingsBody:
      type: object
      properties:
        siteName: { type: string }
        supportWhatsapp: { type: string }
        supportTelegram: { type: string }
        depositUsdtAddress: { type: string }
        depositBankInfo: { type: string }
        depositCashInfo: { type: string }
        minWithdraw: { type: string }
        referralCommissionPct: { type: string }
        referralSignupBonus: { type: string }
        cashbackPct: { type: string }
        maintenanceMode: { type: boolean }
        maintenanceMessage: { type: string }
        announcementBar: { type: string }
        vipBronzeMin: { type: string }
        vipSilverMin: { type: string }
        vipGoldMin: { type: string }
        vipDiamondMin: { type: string }
...
    AdminLoginBody:
      type: object
      required: [password]
      properties:
        username: { type: string }
        password: { type: string }
...
    SupportMessage:
      type: object
      required: [id, sender, authorName, body, createdAt]
      properties:
        id: { type: integer }
        sender: { type: string }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string }
...
    SupportMessageBody:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 4000 }
    AdminSupportTicket:
      type: object
      required: [id, userId, username, subject, category, status, priority, unreadForAdmin, lastUserAt, createdAt, updatedAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        username: { type: string }
        subject: { type: string }
        category: { type: string }
        status: { type: string }
        priority: { type: string }
        unreadForAdmin: { type: integer }
        lastUserAt: { type: string }
        lastAdminAt: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
    AdminSupportTicketDetail:
      type: object
      required: [ticket, messages]
      properties:
        ticket: { $ref: "#/components/schemas/AdminSupportTicket" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/SupportMessage" }
    UpdateAdminSupportTicketBody:
      type: object
      properties:
        status: { type: string }
        priority: { type: string }